Enable user actions when mcm is disabled
